refactor: el servicio que manda a llamar al perfil de usuario se pasó al modulo compartido
- Se implementó el método get_user_profile en el servicio MoodleService para obtener el perfil de un usuario utilizando su ID y el ID del curso.
- Se eliminó el método get_user_profile en MoodleService que estaba duplicado.
- Se actualizó la importación de MoodleService en UserEnrolledCoursesController para utilizar el servicio compartido.
