+++ app/shared/services/moodle_service.py
@@ -36,3 +36,41 @@ class MoodleService:
             user_profile=result["data"][0] if result["data"] else {},
             error=None,
         )
+
+    @staticmethod
+    async def get_user_profile(institute: InstitutesEnum, user_id: int, course_id: int):
+        """
+        Obtener el perfil de un usuario en un curso específico de Moodle.
+        """
+        config = MOODLE_CONFIG[institute]
+        params = {
+            "wstoken": config.moodle_token,
+            "wsfunction": "core_user_get_course_user_profiles",
+            "moodlewsrestformat": "json",
+        }
+
+        data = {
+            "userlist[0][courseid]": course_id,
+            "userlist[0][userid]": user_id,
+        }
+
+        result = await make_moodle_request(
+            url=config.moodle_url,
+            params=params,
+            data=data,
+            institute=institute,
+        )
+
+        if not result["success"]:
+            return SimpleNamespace(
+                user_profile=None,
+                error=result["error_message"],
+            )
+
+        user_profiles = result["data"]
+        user_profile = user_profiles[0] if user_profiles else None
+
+        return SimpleNamespace(
+            user_profile=user_profile,
+            error=None,
+        )
